[638][下线后的基站无法进行删除和编辑操作]

+++ nmp-local/nmp-local-business/business-service/src/main/java/com/matrictime/network/service/impl/BaseStationInfoServiceImpl.java
@@ -148,6 +148,9 @@ public class BaseStationInfoServiceImpl extends SystemBaseService implements Bas
         Result<Integer> result = new Result<>();
         Integer deleteFlag;
         try {
+            if(RequestContext.getUser().getRoleId().equals(DataConstants.SUPER_ADMIN)){
+                throw new SystemException(NO_ADMIN_ERROR_MSG);
+            }
             deleteFlag = baseStationInfoDomainService.deleteBaseStationInfo(baseStationInfoRequest);
             if(deleteFlag.equals(INSERT_OR_UPDATE_SUCCESS)){
                 result.setSuccess(true);
